Kommentare und uprintf (+syscall) hinzugefügt

+++ src/include/threads_handler.h
@@ -3,7 +3,7 @@
 
 
 uint32_t get_running_thread();
-uint32_t load_thread(uint8_t next_thread, uint32_t irq_stackadress);
+uint32_t load_thread(uint32_t next_thread, uint32_t irq_stackadress);
 uint32_t swap_thread(uint32_t irq_stackadress, uint32_t spsr);
 
 #endif //BS_PRAK_THREADS_HANDLER_H
